Convert candle fields to 32-bit types
This speeds up a bit

Wanted to make it more parametrizeable, but meh

Position/Order time is still 64 bit, not that important for performance

+++ src/candle.hpp
@@ -4,10 +4,10 @@
 
 
 struct candle {
-	std::int64_t open_time;
-	double open;
-	double high;
-	double low;
-	double close;
-	double volume;
+	std::uint32_t open_time;
+	float open;
+	float high;
+	float low;
+	float close;
+	float volume;
 };
